requirements.txt
discord.py
discord.py[voice]
requests
PyNaCl
ffmpeg-python
gtts
mysql-connector-python